增加项目业绩字段

+++ applications/personal_center/my_resume.py
@@ -95,11 +95,14 @@ class ResumeDelete(View):
             pk_id = param_dict.get("pk_id", 0)
             if resume_type and pk_id:
                 resume_type_model = resume_model_dict.get(resume_type)
-                deleted, _rows_count = resume_type_model.objects.filter(id=pk_id).delete()
-                if deleted and _rows_count:
-                    result_dict = get_resume_detail_info(custom_user_id)
+                if resume_type_model:
+                    deleted, _rows_count = resume_type_model.objects.filter(id=pk_id).delete()
+                    if deleted and _rows_count:
+                        result_dict = get_resume_detail_info(custom_user_id)
+                    else:
+                        result_dict["msg"] = "未找到要删除的简历信息"
                 else:
-                    result_dict["msg"] = "未找到要删除的简历信息"
+                    result_dict["msg"] = "未找到要删除的简历类型"
             else:
                 result_dict["msg"] = "删除简历信息不完善，删除失败!"
         except:
@@ -131,8 +134,11 @@ class ResumeUpdate(View):
                         if career_objective_obj.exists():
                             resume_info_dict["career_objective"] = career_objective_obj.first()
                 resume_type_model = resume_model_dict.get(resume_type)
-                kwargs = get_kwargs(resume_info_dict)
-                resume_type_model.objects.filter(id=pk_id).update(**kwargs)
+                if resume_type_model:
+                    kwargs = get_kwargs(resume_info_dict)
+                    resume_type_model.objects.filter(id=pk_id).update(**kwargs)
+                else:
+                    result_dict["msg"] = "未找到简历类型"
                 result_dict = get_resume_detail_info(custom_user_id)
             else:
                 result_dict["err"] = 1
@@ -165,21 +171,24 @@ class ResumeAdd(View):
             if user_obj.exists() and resume_info_dict:
                 resume_info_dict["custom_user"] = user_obj.first()
                 resume_type_model = resume_model_dict.get(resume_type)
-                if resume_type == "resume":
-                    resume_obj = Resume.objects.filter(custom_user_id=user_obj)
-                    if resume_obj.exists():
-                        if career_objective_id:
-                            career_objective_obj = CareerObjective.objects.filter(id=career_objective_id)
-                            if career_objective_obj.exists():
-                                resume_info_dict["career_objective"] = career_objective_obj.first()
-                        kwargs = get_kwargs(resume_info_dict)
-                        resume_obj.update(**kwargs)
+                if resume_type_model:
+                    if resume_type == "resume":
+                        resume_obj = Resume.objects.filter(custom_user_id=user_obj)
+                        if resume_obj.exists():
+                            if career_objective_id:
+                                career_objective_obj = CareerObjective.objects.filter(id=career_objective_id)
+                                if career_objective_obj.exists():
+                                    resume_info_dict["career_objective"] = career_objective_obj.first()
+                            kwargs = get_kwargs(resume_info_dict)
+                            resume_obj.update(**kwargs)
+                        else:
+                            kwargs = get_kwargs(resume_info_dict)
+                            resume_type_model.objects.create(**kwargs)
                     else:
                         kwargs = get_kwargs(resume_info_dict)
                         resume_type_model.objects.create(**kwargs)
                 else:
-                    kwargs = get_kwargs(resume_info_dict)
-                    resume_type_model.objects.create(**kwargs)
+                    result_dict["msg"] = "未找到简历类型"
                 result_dict = get_resume_detail_info(custom_user_id)
             else:
                 result_dict["err"] = 1
